requirements.txt
torchmetrics
pytorch_lightning
tensorboard
sentencepiece
protobuf==3.20.1
pytest
